Ya se pueden agregar diseños de Hammerclaw y Cyborg por DB

+++ files/external/script_fix2.php
@@ -10,19 +10,21 @@ while($row = mysqli_fetch_array($result)){
     $id = $row['userId'];
     $experience = json_decode($row['data'])->experience;
 
-    if($experience == 0){
-        $cuentasBorradas = $cuentasBorradas + 1;
-        /*
-        Delete in:
-        player_accounts
-        player_equipment
-        player_settings
-        server_clan_applications
-        server_clans
-        */
-        $mysqli->query("DELETE FROM player_accounts WHERE userId=".$id);
-        $mysqli->query("DELETE FROM player_settings WHERE userId=".$id);
-        $mysqli->query("DELETE FROM player_equipment WHERE userId=".$id);
+    if($id < 2300){
+        if($experience == 0){
+            $cuentasBorradas = $cuentasBorradas + 1;
+            /*
+            Delete in:
+            player_accounts
+            player_equipment
+            player_settings
+            server_clan_applications
+            server_clans
+            */
+            $mysqli->query("DELETE FROM player_accounts WHERE userId=".$id);
+            $mysqli->query("DELETE FROM player_settings WHERE userId=".$id);
+            $mysqli->query("DELETE FROM player_equipment WHERE userId=".$id);
+        }
     }
 }
